feat: add AllowanceManagerClient and related functionality
- Introduced AllowanceManagerClient to handle ERC-20 allowance management.
- Updated SDKAdminManager and SDKManager to include allowance management.
- Created IAllowanceManagerClient interface for client-side allowance operations.
- Implemented server-side handlers for allowance operations including:
  - getApproval
  - isPermit2AuthorizationNeeded
  - getPermit2AuthorizationTx
  - getPermit2RevokeTx
  - getPermit2Data
- Updated intent swap tests to utilize the new allowance management functionality.
- Removed deprecated transaction types and refactored transaction info handling.
