+++ sdk/sdk-client-react/src/handlers/getPermit2UnauthorizeTxHandler.ts
@@ -1,16 +1,15 @@
 import type { ISDKAdminManager, ISDKManager } from '@summerfi/sdk-client'
-import { Address, type AddressValue } from '@summerfi/sdk-common'
+import type { AddressValue, ChainId } from '@summerfi/sdk-common'
 
 /**
  * @name getPermit2RevokeTxHandler
  * @description Creates a transaction to revoke the Permit2 contract authorization for a specific token
+ * @param params.chainId The chain ID where the token lives
  * @param params.tokenAddress The ERC-20 token address to revoke
  * @returns A TransactionInfo for the approve(Permit2, 0) transaction
  */
 export const getPermit2RevokeTxHandler =
   (sdk: ISDKManager | ISDKAdminManager) =>
-  ({ tokenAddress }: { tokenAddress: AddressValue }) => {
-    return sdk.intentSwaps.getPermit2RevokeTx({
-      tokenAddress: Address.createFromEthereum({ value: tokenAddress }),
-    })
+  ({ chainId, tokenAddress }: { chainId: ChainId; tokenAddress: AddressValue }) => {
+    return sdk.allowance.getPermit2RevokeTx({ chainId, tokenAddress })
   }
