+++ sdk/sdk-client/package.json
@@ -40,8 +40,7 @@
     "@summerfi/sdk-common": "workspace:*",
     "@summerfi/sdk-server": "workspace:*",
     "@summerfi/subgraph-manager-common": "workspace:*",
-    "@trpc/client": "11.16.0",
-    "@uniswap/permit2-sdk": "^1.4.0"
+    "@trpc/client": "11.16.0"
   },
   "devDependencies": {
     "@microsoft/api-extractor": "^7.52.13",
