+++ sdk/sdk-client/src/implementation/SDKManager.ts
@@ -11,6 +11,7 @@ import { UsersManager } from './UsersManager'
 import { SimulationManager } from './simulations/SimulationManager'
 import { TokensManagerClient2 } from './TokensManagerClient2'
 import { IntentSwapClient } from './IntentSwapClient'
+import { AllowanceManagerClient } from './AllowanceManagerClient'
 
 /** @see ISDKManager */
 export class SDKManager implements ISDKManager {
@@ -25,6 +26,7 @@ export class SDKManager implements ISDKManager {
   public readonly swaps: SwapManagerClient
   public readonly oracle: OracleManagerClient
   public readonly intentSwaps: IntentSwapClient
+  public readonly allowance: AllowanceManagerClient
 
   public constructor(params: { rpcClient: RPCMainClientType }) {
     this.simulator = new SimulationManager(params)
@@ -38,5 +40,6 @@ export class SDKManager implements ISDKManager {
     this.swaps = new SwapManagerClient(params)
     this.oracle = new OracleManagerClient(params)
     this.intentSwaps = new IntentSwapClient(params)
+    this.allowance = new AllowanceManagerClient(params)
   }
 }
