+++ sdk/sdk-client/src/index.ts
@@ -25,6 +25,7 @@ export type { IUserClient } from './interfaces/IUserClient'
 export type { IUsersManager } from './interfaces/IUsersManager'
 export { PositionUtils } from './utils/PositionUtils'
 export type { CowHook } from './interfaces/IIntentSwapClient'
+export type { IAllowanceManagerClient } from './interfaces/IAllowanceManagerClient'
 export type { UnsignedOrder, EnrichedOrder } from '@cowprotocol/cow-sdk'
 
 export * from './reexports'
