+++ sdk/sdk-client/src/interfaces/ISDKManager.ts
@@ -9,6 +9,7 @@ import type { ITokensManagerClient2 } from './ITokensManagerClient2'
 import type { IUsersManager } from './IUsersManager'
 import type { ISimulationManager } from './simulations/ISimulationManager'
 import type { IIntentSwapClient } from './IIntentSwapClient'
+import type { IAllowanceManagerClient } from './IAllowanceManagerClient'
 
 /**
  * SDKManager is the main entry point for interacting with the SDK in the client side
@@ -38,4 +39,6 @@ export interface ISDKManager {
   readonly oracle: IOracleManagerClient
   /** Intent Swap Client for interacting with CoW Protocol intent swaps */
   readonly intentSwaps: IIntentSwapClient
+  /** Allowance Manager Client for Permit2 authorization checks, transactions and typed data */
+  readonly allowance: IAllowanceManagerClient
 }
