+++ sdk/sdk-common/src/common/types/Permit2PermitData.ts
@@ -1,0 +1,11 @@
+import type { AddressValue } from './AddressValue'
+
+/**
+ * @name Permit2PermitData
+ * @description The PermitTransferFrom payload consumed by AdmiralsQuarters permit2 calldata
+ */
+export type Permit2PermitData = {
+  permitted: { token: AddressValue; amount: bigint }
+  nonce: bigint
+  deadline: bigint
+}
