+++ sdk/sdk-server/src/handlers/allowanceGetPermit2Data.ts
@@ -1,0 +1,17 @@
+import { z } from 'zod'
+import { publicProcedure } from '../SDKTRPC'
+import { isAddressValue, isChainId, type AddressValue, type ChainId } from '@summerfi/sdk-common'
+
+export const allowanceGetPermit2Data = publicProcedure
+  .input(
+    z.object({
+      chainId: z.custom<ChainId>(isChainId),
+      tokenAddress: z.custom<AddressValue>(isAddressValue),
+      amount: z.bigint(),
+      spenderAddress: z.custom<AddressValue>(isAddressValue),
+      senderAddress: z.custom<AddressValue>(isAddressValue),
+    }),
+  )
+  .query(async (opts) => {
+    return opts.ctx.allowanceManager.getPermit2Data(opts.input)
+  })
